Create pkg/api/util and re-arrange controller instantiation code
