+++ pkg/apis/certmanager/v1alpha1/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
         "conversion.go",
         "defaults.go",
         "doc.go",
-        "helpers.go",
+        "generic_issuer.go",
         "register.go",
         "types.go",
         "types_certificate.go",
@@ -21,7 +21,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/certmanager:go_default_library",
-        "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
