+++ pkg/apis/certmanager/v1alpha1/generic_issuer.go
@@ -1,0 +1,106 @@
+/*
+Copyright 2019 The Jetstack cert-manager contributors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+import (
+	"fmt"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	runtime "k8s.io/apimachinery/pkg/runtime"
+)
+
+type GenericIssuer interface {
+	runtime.Object
+	GetObjectMeta() *metav1.ObjectMeta
+	GetSpec() *IssuerSpec
+	GetStatus() *IssuerStatus
+}
+
+var _ GenericIssuer = &Issuer{}
+var _ GenericIssuer = &ClusterIssuer{}
+
+func (c *ClusterIssuer) GetObjectMeta() *metav1.ObjectMeta {
+	return &c.ObjectMeta
+}
+func (c *ClusterIssuer) GetSpec() *IssuerSpec {
+	return &c.Spec
+}
+func (c *ClusterIssuer) GetStatus() *IssuerStatus {
+	return &c.Status
+}
+func (c *ClusterIssuer) SetSpec(spec IssuerSpec) {
+	c.Spec = spec
+}
+func (c *ClusterIssuer) SetStatus(status IssuerStatus) {
+	c.Status = status
+}
+func (c *ClusterIssuer) Copy() GenericIssuer {
+	return c.DeepCopy()
+}
+func (c *Issuer) GetObjectMeta() *metav1.ObjectMeta {
+	return &c.ObjectMeta
+}
+func (c *Issuer) GetSpec() *IssuerSpec {
+	return &c.Spec
+}
+func (c *Issuer) GetStatus() *IssuerStatus {
+	return &c.Status
+}
+func (c *Issuer) SetSpec(spec IssuerSpec) {
+	c.Spec = spec
+}
+func (c *Issuer) SetStatus(status IssuerStatus) {
+	c.Status = status
+}
+func (c *Issuer) Copy() GenericIssuer {
+	return c.DeepCopy()
+}
+
+// TODO: refactor these functions away
+func (i *IssuerStatus) ACMEStatus() *ACMEIssuerStatus {
+	// this is an edge case, but this will prevent panics
+	if i == nil {
+		return &ACMEIssuerStatus{}
+	}
+	if i.ACME == nil {
+		i.ACME = &ACMEIssuerStatus{}
+	}
+	return i.ACME
+}
+
+func (a *ACMEIssuerDNS01Config) Provider(name string) (*ACMEIssuerDNS01Provider, error) {
+	if a == nil {
+		return nil, fmt.Errorf("issuer does not contain DNS01 configuration for provider named %q", name)
+	}
+	for _, p := range a.Providers {
+		if p.Name == name {
+			return &(*&p), nil
+		}
+	}
+	return nil, fmt.Errorf("issuer does not contain DNS01 configuration for provider named %q", name)
+}
+
+func ConfigForDomain(cfgs []DomainSolverConfig, domain string) *DomainSolverConfig {
+	for _, cfg := range cfgs {
+		for _, d := range cfg.Domains {
+			if d == domain {
+				return &cfg
+			}
+		}
+	}
+	return &DomainSolverConfig{}
+}
