+++ pkg/apis/certmanager/validation/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/apis/certmanager/validation",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
-        "//pkg/controller:go_default_library",
         "//pkg/issuer/acme/dns/rfc2136:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/validation/field:go_default_library",
