+++ pkg/controller/BUILD.bazel
@@ -5,7 +5,6 @@ go_library(
     srcs = [
         "context.go",
         "helper.go",
-        "issuer_factory.go",
         "register.go",
         "util.go",
     ],
@@ -15,8 +14,6 @@ go_library(
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
         "//pkg/client/clientset/versioned:go_default_library",
         "//pkg/client/informers/externalversions:go_default_library",
-        "//pkg/client/listers/certmanager/v1alpha1:go_default_library",
-        "//pkg/issuer:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/runtime:go_default_library",
