+++ pkg/controller/acmeorders/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
         "//pkg/client/listers/certmanager/v1alpha1:go_default_library",
         "//pkg/controller:go_default_library",
+        "//pkg/issuer:go_default_library",
         "//pkg/util:go_default_library",
         "//third_party/crypto/acme:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
