+++ pkg/controller/issuers/controller.go
@@ -33,11 +33,13 @@ import (
 
 	cmlisters "github.com/jetstack/cert-manager/pkg/client/listers/certmanager/v1alpha1"
 	controllerpkg "github.com/jetstack/cert-manager/pkg/controller"
+	"github.com/jetstack/cert-manager/pkg/issuer"
 	"github.com/jetstack/cert-manager/pkg/util"
 )
 
 type Controller struct {
 	*controllerpkg.Context
+	issuerFactory issuer.IssuerFactory
 
 	// To allow injection for testing.
 	syncHandler func(ctx context.Context, key string) error
@@ -66,6 +68,7 @@ func New(ctx *controllerpkg.Context) *Controller {
 	secretsInformer.Informer().AddEventHandler(&controllerpkg.BlockingEventHandler{WorkFunc: ctrl.secretDeleted})
 	ctrl.watchedInformers = append(ctrl.watchedInformers, secretsInformer.Informer().HasSynced)
 	ctrl.secretLister = secretsInformer.Lister()
+	ctrl.issuerFactory = issuer.NewIssuerFactory(ctx)
 
 	return ctrl
 }
