+++ pkg/controller/issuers/sync.go
@@ -25,6 +25,7 @@ import (
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/errors"
 
+	apiutil "github.com/jetstack/cert-manager/pkg/api/util"
 	"github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
 	"github.com/jetstack/cert-manager/pkg/apis/certmanager/validation"
 )
@@ -47,10 +48,11 @@ func (c *Controller) Sync(ctx context.Context, iss *v1alpha1.Issuer) (err error)
 	el := validation.ValidateIssuer(issuerCopy)
 	if len(el) > 0 {
 		msg := fmt.Sprintf("Resource validation failed: %v", el.ToAggregate())
-		issuerCopy.UpdateStatusCondition(v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorConfig, msg)
+		apiutil.SetIssuerCondition(issuerCopy, v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorConfig, msg)
 		return
 	}
 
+	// Remove existing ErrorConfig condition if it exists
 	for i, c := range issuerCopy.Status.Conditions {
 		if c.Type == v1alpha1.IssuerConditionReady {
 			if c.Reason == errorConfig && c.Status == v1alpha1.ConditionFalse {
@@ -60,7 +62,7 @@ func (c *Controller) Sync(ctx context.Context, iss *v1alpha1.Issuer) (err error)
 		}
 	}
 
-	i, err := c.IssuerFactory().IssuerFor(issuerCopy)
+	i, err := c.issuerFactory.IssuerFor(issuerCopy)
 
 	if err != nil {
 		return err
