+++ pkg/controller/register.go
@@ -16,13 +16,6 @@ limitations under the License.
 
 package controller
 
-import (
-	"sync"
-
-	"github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
-	"github.com/jetstack/cert-manager/pkg/issuer"
-)
-
 // This file defines types for controllers to register themselves with the
 // controller package.
 
@@ -49,24 +42,4 @@ func Known() map[string]Constructor {
 // Register registers a controller constructor with the controller package
 func Register(name string, fn Constructor) {
 	known[name] = fn
-}
-
-// issuerConstructor constructs an issuer given an Issuer resource and a Context.
-// An error will be returned if the appropriate issuer is not registered.
-type IssuerConstructor func(*Context, v1alpha1.GenericIssuer) (issuer.Interface, error)
-
-var (
-	constructors     = make(map[string]IssuerConstructor)
-	constructorsLock sync.RWMutex
-)
-
-// Register will register an issuer constructor so it can be used within the
-// application. 'name' should be unique, and should be used to identify this
-// issuer.
-// TODO: move this method to be on Factory, and invent a way to obtain a
-// SharedFactory. This will make testing easier.
-func RegisterIssuer(name string, c IssuerConstructor) {
-	constructorsLock.Lock()
-	defer constructorsLock.Unlock()
-	constructors[name] = c
 }
