+++ pkg/issuer/acme/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
     deps = [
         "//pkg/acme:go_default_library",
         "//pkg/acme/client:go_default_library",
+        "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
         "//pkg/client/listers/certmanager/v1alpha1:go_default_library",
         "//pkg/controller:go_default_library",
