+++ pkg/issuer/ca/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/issuer/ca",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/issuer:go_default_library",
