+++ pkg/issuer/ca/setup.go
@@ -22,6 +22,7 @@ import (
 	"github.com/golang/glog"
 	"k8s.io/api/core/v1"
 
+	apiutil "github.com/jetstack/cert-manager/pkg/api/util"
 	"github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
 	"github.com/jetstack/cert-manager/pkg/util/kube"
 )
@@ -44,7 +45,7 @@ func (c *CA) Setup(ctx context.Context) error {
 		s := messageErrorGetKeyPair + err.Error()
 		glog.Info(s)
 		c.Recorder.Event(c.issuer, v1.EventTypeWarning, errorGetKeyPair, s)
-		c.issuer.UpdateStatusCondition(v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorGetKeyPair, s)
+		apiutil.SetIssuerCondition(c.issuer, v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorGetKeyPair, s)
 		return err
 	}
 
@@ -53,7 +54,7 @@ func (c *CA) Setup(ctx context.Context) error {
 		s := messageErrorGetKeyPair + err.Error()
 		glog.Info(s)
 		c.Recorder.Event(c.issuer, v1.EventTypeWarning, errorGetKeyPair, s)
-		c.issuer.UpdateStatusCondition(v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorGetKeyPair, s)
+		apiutil.SetIssuerCondition(c.issuer, v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorGetKeyPair, s)
 		return err
 	}
 
@@ -61,14 +62,14 @@ func (c *CA) Setup(ctx context.Context) error {
 		s := messageErrorGetKeyPair + "certificate is not a CA"
 		glog.Info(s)
 		c.Recorder.Event(c.issuer, v1.EventTypeWarning, errorInvalidKeyPair, s)
-		c.issuer.UpdateStatusCondition(v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorInvalidKeyPair, s)
+		apiutil.SetIssuerCondition(c.issuer, v1alpha1.IssuerConditionReady, v1alpha1.ConditionFalse, errorInvalidKeyPair, s)
 		// Don't return an error here as there is nothing more we can do
 		return nil
 	}
 
 	glog.Info(messageKeyPairVerified)
 	c.Recorder.Event(c.issuer, v1.EventTypeNormal, successKeyPairVerified, messageKeyPairVerified)
-	c.issuer.UpdateStatusCondition(v1alpha1.IssuerConditionReady, v1alpha1.ConditionTrue, successKeyPairVerified, messageKeyPairVerified)
+	apiutil.SetIssuerCondition(c.issuer, v1alpha1.IssuerConditionReady, v1alpha1.ConditionTrue, successKeyPairVerified, messageKeyPairVerified)
 
 	return nil
 }
